+++ lib/src/iso/bmff/boxes.ts
@@ -5,16 +5,21 @@ export type { Box } from './boxes/Box.js';
 export type { ChunkLargeOffsetBox } from './boxes/ChunkLargeOffsetBox.js';
 export type { ChunkOffsetBox } from './boxes/ChunkOffsetBox.js';
 export type { CompactSampleSizeBox } from './boxes/CompactSampleSizeBox.js';
-export type { CompositionTimeToSampleBox, CompositionTimeToSampleEntry } from './boxes/CompositionTimeToSampleBox.js';
+export type { CompositionTimeToSampleBox } from './boxes/CompositionTimeToSampleBox.js';
+export type { CompositionTimeToSampleEntry } from './boxes/CompositionTimeToSampleEntry.js';
 export type { ContainerBox } from './boxes/ContainerBox.js';
 export type { DataEntryUrlBox } from './boxes/DataEntryUrlBox.js';
 export type { DataEntryUrnBox } from './boxes/DataEntryUrnBox.js';
 export type { DataInformationBox } from './boxes/DataInformationBox.js';
 export type { DataReferenceBox } from './boxes/DataReferenceBox.js';
-export type { DecodingTimeSample, DecodingTimeToSampleBox } from './boxes/DecodingTimeToSampleBox.js';
+export type { DecodingTimeSample } from './boxes/DecodingTimeSample.js';
+export type { DecodingTimeToSampleBox } from './boxes/DecodingTimeToSampleBox.js';
 export type { DegradationPriorityBox } from './boxes/DegradationPriorityBox.js';
 export type { EditBox } from './boxes/EditBox.js';
-export type { EditListBox, EditListEntry } from './boxes/EditListBox.js';
+export type { EditListBox } from './boxes/EditListBox.js';
+export type { EditListEntry } from './boxes/EditListEntry.js';
+export type { EncryptedSample } from './boxes/EncryptedSample.js';
+export type { Entity } from './boxes/Entity.js';
 export type { EventMessageBox } from './boxes/EventMessageBox.js';
 export type { ExtendedLanguageBox } from './boxes/ExtendedLanguageBox.js';
 export type { Fields } from './boxes/Fields.js';
@@ -51,7 +56,7 @@ export type { MovieFragmentRandomAccessOffsetBox } from './boxes/MovieFragmentRa
 export type { MovieHeaderBox } from './boxes/MovieHeaderBox.js';
 export type { NullMediaHeaderBox } from './boxes/NullMediaHeaderBox.js';
 export type { OriginalFormatBox } from './boxes/OriginalFormatBox.js';
-export type { Entity, PreselectionGroupBox } from './boxes/PreselectionGroupBox.js';
+export type { PreselectionGroupBox } from './boxes/PreselectionGroupBox.js';
 export type { PrimaryItemBox } from './boxes/PrimaryItemBox.js';
 export type { ProducerReferenceTimeBox } from './boxes/ProducerReferenceTimeBox.js';
 export type { ProtectionSchemeInformationBox } from './boxes/ProtectionSchemeInformationBox.js';
@@ -60,7 +65,7 @@ export type { SampleAuxiliaryInformationOffsetsBox } from './boxes/SampleAuxilia
 export type { SampleAuxiliaryInformationSizesBox } from './boxes/SampleAuxiliaryInformationSizesBox.js';
 export type { SampleDependencyTypeBox } from './boxes/SampleDependencyTypeBox.js';
 export type { SampleDescriptionBox } from './boxes/SampleDescriptionBox.js';
-export type { EncryptedSample, SampleEncryptionBox, SubsampleEncryption } from './boxes/SampleEncryptionBox.js';
+export type { SampleEncryptionBox } from './boxes/SampleEncryptionBox.js';
 export type { SampleEntryBox } from './boxes/SampleEntryBox.js';
 export type { SampleGroupDescriptionBox } from './boxes/SampleGroupDescriptionBox.js';
 export type { SampleSizeBox } from './boxes/SampleSizeBox.js';
@@ -71,16 +76,23 @@ export type { SampleToGroupBox } from './boxes/SampleToGroupBox.js';
 export type { SampleToGroupEntry } from './boxes/SampleToGroupEntry.js';
 export type { SchemeInformationBox } from './boxes/SchemeInformationBox.js';
 export type { SchemeTypeBox } from './boxes/SchemeTypeBox.js';
-export type { SegmentIndexBox, SegmentIndexReference } from './boxes/SegmentIndexBox.js';
+export type { SegmentIndexBox } from './boxes/SegmentIndexBox.js';
+export type { SegmentIndexReference } from './boxes/SegmentIndexReference.js';
 export type { SegmentTypeBox } from './boxes/SegmentTypeBox.js';
 export type { ShadowSyncEntry } from './boxes/ShadowSyncEntry.js';
 export type { ShadowSyncSampleBox } from './boxes/ShadowSyncSampleBox.js';
 export type { SingleItemTypeReferenceBox } from './boxes/SingleItemTypeReferenceBox.js';
 export type { SoundMediaHeaderBox } from './boxes/SoundMediaHeaderBox.js';
-export type { Subsample, SubsampleEntry, SubsampleInformationBox } from './boxes/SubsampleInformationBox.js';
-export type { Range, Subsegment, SubsegmentIndexBox } from './boxes/SubsegmentIndexBox.js';
+export type { Subsample } from './boxes/Subsample.js';
+export type { SubsampleEncryption } from './boxes/SubsampleEncryption.js';
+export type { SubsampleEntry } from './boxes/SubsampleEntry.js';
+export type { SubsampleInformationBox } from './boxes/SubsampleInformationBox.js';
+export type { Subsegment } from './boxes/Subsegment.js';
+export type { SubsegmentIndexBox } from './boxes/SubsegmentIndexBox.js';
+export type { SubsegmentRange } from './boxes/SubsegmentRange.js';
 export type { SubtitleMediaHeaderBox } from './boxes/SubtitleMediaHeaderBox.js';
-export type { SyncSample, SyncSampleBox } from './boxes/SyncSampleBox.js';
+export type { SyncSample } from './boxes/SyncSample.js';
+export type { SyncSampleBox } from './boxes/SyncSampleBox.js';
 export type { TrackBox } from './boxes/TrackBox.js';
 export type { TrackEncryptionBox } from './boxes/TrackEncryptionBox.js';
 export type { TrackExtendsBox } from './boxes/TrackExtendsBox.js';
