+++ lib/src/iso/bmff/boxes/CompositionTimeToSampleBox.ts
@@ -1,16 +1,5 @@
+import type { CompositionTimeToSampleEntry } from './CompositionTimeToSampleEntry.js';
 import type { FullBox } from './FullBox.js';
-
-/**
- * A Composition Time To Sample Entry
- *
- * @group ISOBMFF
- *
- * @beta
- */
-export type CompositionTimeToSampleEntry = {
-	sampleCount: number;
-	sampleOffset: number;
-};
 
 /**
  * ISO/IEC 14496-12:2012 - 8.6.1.3 Composition Time To Sample Box
