+++ lib/src/iso/bmff/boxes/CompositionTimeToSampleEntry.ts
@@ -1,0 +1,11 @@
+/**
+ * A Composition Time To Sample Entry
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type CompositionTimeToSampleEntry = {
+	sampleCount: number;
+	sampleOffset: number;
+};
