+++ lib/src/iso/bmff/boxes/DecodingTimeToSampleBox.ts
@@ -1,19 +1,5 @@
+import type { DecodingTimeSample } from './DecodingTimeSample.js';
 import type { FullBox } from './FullBox.js';
-
-/**
- * sample
- *
- * @group ISOBMFF
- *
- * @beta
- */
-export type DecodingTimeSample = {
-	/** A 32-bit integer that specifies the number of consecutive samples that have the same decoding time delta. */
-	sampleCount: number;
-
-	/** A 32-bit integer that specifies the delta of the decoding time of each sample in the table. */
-	sampleDelta: number;
-};
 
 /**
  * ISO/IEC 14496-12:2012 - 8.6.1.2 Decoding Time To Sample Box
