+++ lib/src/iso/bmff/boxes/EditListBox.ts
@@ -1,18 +1,5 @@
+import type { EditListEntry } from './EditListEntry.js';
 import type { FullBox } from './FullBox.js';
-
-/**
- * An edit list entry.
- *
- * @group ISOBMFF
- *
- * @beta
- */
-export type EditListEntry = {
-	segmentDuration: number;
-	mediaTime: number;
-	mediaRateInteger: number;
-	mediaRateFraction: number;
-};
 
 /**
  * ISO/IEC 14496-12:2012 - 8.6.6 Edit List Box
