+++ lib/src/iso/bmff/boxes/EditListEntry.ts
@@ -1,0 +1,13 @@
+/**
+ * An edit list entry.
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type EditListEntry = {
+	segmentDuration: number;
+	mediaTime: number;
+	mediaRateInteger: number;
+	mediaRateFraction: number;
+};
