+++ lib/src/iso/bmff/boxes/EncryptedSample.ts
@@ -1,0 +1,13 @@
+import type { SubsampleEncryption } from './SubsampleEncryption.js';
+
+/**
+ * Encrypted Sample
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type EncryptedSample = {
+	initializationVector?: Uint8Array;
+	subsampleEncryption?: SubsampleEncryption[];
+};
