+++ lib/src/iso/bmff/boxes/Entity.ts
@@ -1,0 +1,11 @@
+/**
+ * Entity
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type Entity = {
+	/** Entity ID */
+	entityId: number;
+};
