+++ lib/src/iso/bmff/boxes/SegmentIndexReference.ts
@@ -1,0 +1,17 @@
+/**
+ * Segment index reference
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SegmentIndexReference = {
+	reference: number;
+	subsegmentDuration: number;
+	sap: number;
+	referenceType: number;
+	referencedSize: number;
+	startsWithSap: number;
+	sapType: number;
+	sapDeltaTime: number;
+};
