+++ lib/src/iso/bmff/boxes/Subsample.ts
@@ -1,0 +1,13 @@
+/**
+ * Sub sample
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type Subsample = {
+	subsampleSize: number;
+	subsamplePriority: number;
+	discardable: number;
+	codecSpecificParameters: number;
+};
