+++ lib/src/iso/bmff/boxes/SubsampleEncryption.ts
@@ -1,0 +1,11 @@
+/**
+ * Subsample Encryption
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SubsampleEncryption = {
+	bytesOfClearData: number;
+	bytesOfProtectedData: number;
+};
