+++ lib/src/iso/bmff/boxes/SubsampleEntry.ts
@@ -1,0 +1,14 @@
+import type { Subsample } from './Subsample.js';
+
+/**
+ * Sub sample entry
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SubsampleEntry = {
+	sampleDelta: number;
+	subsampleCount: number;
+	subsamples: Subsample[];
+};
