+++ lib/src/iso/bmff/boxes/Subsegment.ts
@@ -1,0 +1,13 @@
+import type { SubsegmentRange } from './SubsegmentRange.js';
+
+/**
+ * Subsegment
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type Subsegment = {
+	rangesCount: number;
+	ranges: SubsegmentRange[];
+};
