+++ lib/src/iso/bmff/boxes/SubsegmentRange.ts
@@ -1,0 +1,11 @@
+/**
+ * Subsegment range
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SubsegmentRange = {
+	level: number;
+	rangeSize: number;
+};
