+++ lib/src/iso/bmff/boxes/SyncSample.ts
@@ -1,0 +1,10 @@
+/**
+ * Sync sample
+ *
+ * @group ISOBMFF
+ *
+ * @beta
+ */
+export type SyncSample = {
+	sampleNumber: number;
+};
